postpractical_modifications
Added the markdown and html files for the postpractical study. Answers of the questions and practices are includes. In addition, some extra plots to visualise PC related genes as well as information on tSNE and UMAP are included. These are meant to be informative, are very advanced and not part fo the course/exam. Finally, the markdown folder is disbanded; the Rmd files are now on the main folder